requirements.txt
pyee==8.1.0
mplfinance==0.12.7a17
pandas~=1.3.2
python-dotenv==0.19.0
scipy==1.7.1
ta==0.7.0
matplotlib==3.4.2
requests==2.26.0
APScheduler==3.6.3
websocket-client==1.1.1
setuptools==57.4.0
python-telegram-bot==13.7
SQLAlchemy==1.4.22
telegram==0.0.1
tzlocal==2.1
pytz~=2021.1
fastapi~=0.68.0
pydantic~=1.8.2
uvicorn~=0.15.0